popcorn: Renamed arch_exception_state_t to arch_context_t and updated arch_fast_move and arch_fast_zero function prototype

+++ kernel/vfs/vfs_node.c
@@ -31,8 +31,8 @@ void vfs_node_init(void) {
     kmem_slab_create(&vfs_node_slab, sizeof(vfs_node_t), VFS_NODE_SLAB_NUM);
 
     // Initialize the hash table
-    arch_fast_zero((uintptr_t)vfs_node_hash_table.ll_vn_buckets, NUM_BUCKETS * sizeof(list_t));
-    arch_fast_zero((uintptr_t)vfs_node_hash_table.bkt_lock, NUM_BUCKETS * sizeof(spinlock_t));
+    arch_fast_zero(vfs_node_hash_table.ll_vn_buckets, NUM_BUCKETS * sizeof(list_t));
+    arch_fast_zero(vfs_node_hash_table.bkt_lock, NUM_BUCKETS * sizeof(spinlock_t));
 }
 
 vfs_node_t* vfs_node_get(struct vfs_mount_s *mnt, vfs_ino_t id) {
